✦ Refactoring snapshot swr
